+++ src/time-picker/index.ts
@@ -3,3 +3,5 @@ export * from './inner/inner-time-picker.config';
 export * from './time-picker.module';
 export * from './inner/inner-time-picker.store';
 export * from './time-picker.utils';
+export * from './time-picker.component';
+export * from './time-picker-panel.component';